move game logic to its plugin
