+++ client/src/interface/plugin.rs
@@ -3,7 +3,6 @@ use std::str::FromStr;
 
 use bevy::app::{App, AppExit, Plugin, Update};
 use bevy::asset::AssetServer;
-use bevy::audio::AudioBundle;
 use bevy::ecs::change_detection::{Res, ResMut};
 use bevy::ecs::entity::Entity;
 use bevy::ecs::event::EventWriter;
@@ -20,14 +19,12 @@ use bevy::ui::widget::Button;
 use bevy::ui::Interaction;
 use bevy::utils::default;
 use bevy_simple_text_input::{TextInputInactive, TextInputPlugin, TextInputValue};
-use game_server::game::game::{BoardCell, FinishedState, GameState};
+use game_server::game::game::{FinishedState, GameState};
 
 use crate::app_state::{AppState, AppStateTransition, MenuState};
 use crate::board;
-use crate::game::{
-    CellUpdated, CurrentGame, FullGameInfo, GameInfo, GameOver, GamePlugin, MockBotGame, Position,
-    StateUpdated,
-};
+use crate::commands::CommandsExt;
+use crate::game::{CellUpdated, CurrentGame, GameInfo, GameOver, GamePlugin, StateUpdated};
 use crate::grpc::{CallCreateGame, CallGetGame, CallGetPlayerGames, CallMakeTurn, GrpcClient};
 use crate::interface::buttons::{
     spawn_exit_button, spawn_join_game_button_bundle, spawn_menu_navigation_button, JoinGame,
@@ -38,20 +35,12 @@ use crate::interface::common::button_bundle::{
 use crate::interface::common::{
     column_node_bundle, global_column_node_bundle, menu_item_style, menu_text_bundle,
     menu_text_input_bundle, menu_text_style, row_node_bundle, CONFIRMATION_SOUND_PATH,
-    ERROR_SOUND_PATH, GAME_LIST_REFRESH_INTERVAL_SEC, O_SPRITE_PATH, TURN_SOUND_PATH,
-    X_SPRITE_PATH,
+    ERROR_SOUND_PATH, GAME_LIST_REFRESH_INTERVAL_SEC,
 };
 use crate::interface::components::{overlay_ui_node, AssociatedTextInput, Overlay};
 use crate::interface::game_list::{GameList, GameListBundle, LoadingGameListBundle};
 use crate::interface::ingame::{InGameUIBundle, InGameUIPlugin, PlayerInfoReady};
 use crate::settings::{Settings, SubmitTextInputSetting};
-
-fn play_sound(commands: &mut Commands, asset_server: &AssetServer, sound_path: &'static str) {
-    commands.spawn(AudioBundle {
-        source: asset_server.load(sound_path),
-        ..default()
-    });
-}
 
 #[derive(Deref, DerefMut, Resource)]
 struct RefreshGamesTimer(pub Timer);
@@ -110,19 +99,8 @@ impl Plugin for InterfacePlugin {
                     (refresh_game_list, handle_player_games_task, join_game)
                         .run_if(in_state(AppState::Menu(MenuState::PlayOverNetwork))),
                     handle_create_game_task,
-                    handle_make_turn_task,
-                    handle_get_game_task,
-                    (
-                        make_turn,
-                        refresh_game_board.run_if(not(any_with_component::<CallGetGame>)),
-                        handle_cell_updated,
-                        handle_state_updated,
-                        handle_game_over.after(handle_state_updated),
-                    )
+                    (make_turn, handle_cell_updated, handle_game_over)
                         .run_if(in_state(AppState::Game).and_then(resource_exists::<CurrentGame>)),
-                    create_mock_bot_game
-                        .after(state_transition)
-                        .run_if(in_state(AppState::Menu(MenuState::PlayWithBot))),
                 ),
             );
     }
@@ -141,7 +119,6 @@ fn state_transition(
     mut commands: Commands,
     mut next_app_state: ResMut<NextState<AppState>>,
     mut exit: EventWriter<AppExit>,
-    mut mock_bot_game: EventWriter<MockBotGame>,
     app_state: Res<State<AppState>>,
     settings: Res<Settings>,
     keyboard_input: Res<ButtonInput<KeyCode>>,
@@ -174,18 +151,27 @@ fn state_transition(
                                 }
                             }
                         } else {
-                            play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                            commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                         }
                     }
                 } else if *app_state.get() == AppState::Menu(MenuState::PlayWithBot)
                     && new_state == AppState::Game
                 {
                     if let Some(user_id) = settings.user_id() {
-                        mock_bot_game.send(MockBotGame { user_id });
+                        let game = CurrentGame::new_from_asset_server(
+                            user_id,
+                            GameInfo {
+                                id: 0,
+                                state: GameState::Finished(FinishedState::Draw),
+                                players: vec![user_id, u64::MAX], // TODO: replace this with bot id
+                            },
+                            &asset_server,
+                        );
+                        commands.insert_resource(game);
                         println!("state transition: {:?}", new_state);
                         next_app_state.set(new_state);
                     } else {
-                        play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                        commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                     }
                 } else {
                     // regular state transition
@@ -197,30 +183,6 @@ fn state_transition(
                 exit.send(AppExit);
             }
         }
-    }
-}
-
-fn create_mock_bot_game(
-    mut commands: Commands,
-    mut game_over: EventWriter<StateUpdated>,
-    mut mock_bot_game: EventReader<MockBotGame>,
-    asset_server: Res<AssetServer>,
-) {
-    if let Some(&MockBotGame { user_id }) = mock_bot_game.read().last() {
-        let x_img = asset_server.load(X_SPRITE_PATH);
-        let o_img = asset_server.load(O_SPRITE_PATH);
-        let game = CurrentGame::new(
-            user_id,
-            GameInfo {
-                id: 0,
-                state: GameState::Finished(FinishedState::Draw),
-                players: vec![user_id, u64::MAX], // TODO: replace this with bot id
-            },
-            x_img,
-            o_img,
-        );
-        game_over.send(StateUpdated(game.state()));
-        commands.insert_resource(game);
     }
 }
 
@@ -240,13 +202,10 @@ fn join_game(
         }
     };
     if let Some((_, join)) = join_button.iter().find(|(&i, _)| i == Interaction::Pressed) {
-        let x_img = asset_server.load(X_SPRITE_PATH);
-        let o_img = asset_server.load(O_SPRITE_PATH);
-        commands.insert_resource(CurrentGame::new(
+        commands.insert_resource(CurrentGame::new_from_asset_server(
             user_id,
             join.deref().clone(),
-            x_img,
-            o_img,
+            &asset_server,
         ));
         if let Some(task) = grpc_client.load_game(join.id) {
             commands.spawn(CallGetGame(task));
@@ -255,7 +214,7 @@ fn join_game(
         }
         println!("state transition: join game");
         next_app_state.set(AppState::Game);
-        play_sound(&mut commands, &asset_server, CONFIRMATION_SOUND_PATH);
+        commands.play_sound(&asset_server, CONFIRMATION_SOUND_PATH);
     }
 }
 
@@ -288,9 +247,9 @@ fn settings_submit<T: FromStr + 'static>(
             {
                 if let Ok(val) = val.0.parse::<T>() {
                     submit_input.submit(&mut settings, val);
-                    play_sound(&mut commands, &asset_server, CONFIRMATION_SOUND_PATH);
+                    commands.play_sound(&asset_server, CONFIRMATION_SOUND_PATH);
                 } else {
-                    play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                    commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                 }
             }
         }
@@ -550,7 +509,7 @@ fn make_turn(
             commands.spawn((CallMakeTurn(task), event.pos));
         } else {
             println!("grpc server is down");
-            play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+            commands.play_sound(&asset_server, ERROR_SOUND_PATH);
         }
     }
 }
@@ -586,20 +545,6 @@ fn refresh_game_list(
                         parent.spawn(menu_text_bundle("No user id provided", text_style.clone()));
                     });
             }
-        }
-    }
-}
-
-fn refresh_game_board(
-    mut commands: Commands,
-    mut timer: ResMut<RefreshGamesTimer>,
-    game: Res<CurrentGame>,
-    client: ResMut<GrpcClient>,
-    time: Res<Time>,
-) {
-    if timer.tick(time.delta()).just_finished() {
-        if let Some(task) = client.load_game(game.id()) {
-            commands.spawn(CallGetGame(task));
         }
     }
 }
@@ -705,7 +650,7 @@ fn handle_create_game_task(
                 Ok(response) => {
                     let Some(game_info) = response.into_inner().game_info else {
                         println!("received empty CreateGame reply from server");
-                        play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                        commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                         return;
                     };
                     let info = match GameInfo::try_from(game_info) {
@@ -717,135 +662,23 @@ fn handle_create_game_task(
                     };
                     if let Some(&user_id) = info.players.first() {
                         println!("starting created game: {}", info.id);
-                        let x_img = asset_server.load(X_SPRITE_PATH);
-                        let o_img = asset_server.load(O_SPRITE_PATH);
-                        commands.insert_resource(CurrentGame::new(user_id, info, x_img, o_img));
+                        commands.insert_resource(CurrentGame::new_from_asset_server(
+                            user_id,
+                            info,
+                            &asset_server,
+                        ));
                         next_app_state.set(AppState::Game);
-                        play_sound(&mut commands, &asset_server, CONFIRMATION_SOUND_PATH);
+                        commands.play_sound(&asset_server, CONFIRMATION_SOUND_PATH);
                     } else {
                         println!("GameInfo is corrupted: players is empty");
-                        play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                        commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                     }
                 }
                 Err(err) => {
                     println!("CreateGame request failed: {}", err);
-                    play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
+                    commands.play_sound(&asset_server, ERROR_SOUND_PATH);
                 }
             };
-        }
-    }
-}
-
-fn handle_make_turn_task(
-    mut make_turn: Query<(Entity, &mut CallMakeTurn, &Position)>,
-    mut commands: Commands,
-    mut game: Option<ResMut<CurrentGame>>,
-    mut cell_updated: EventWriter<CellUpdated>,
-    mut state_updated: EventWriter<StateUpdated>,
-    asset_server: Res<AssetServer>,
-) {
-    for (entity, mut task, pos) in make_turn.iter_mut() {
-        if let Some(res) = block_on(future::poll_once(&mut task.0)) {
-            commands.entity(entity).despawn();
-            let game = match &mut game {
-                Some(game) => game,
-                None => {
-                    println!("no current game, dropping MakeTurn response");
-                    continue;
-                }
-            };
-            let user_id = game.user_id();
-            match res {
-                Ok(response) => {
-                    if let Some(next) = game.get_next_player() {
-                        if next != user_id {
-                            println!("state is already updated, skip handling MakeTurn response");
-                            continue;
-                        }
-                    }
-                    if let Some(new_state) = response.into_inner().game_state {
-                        let state = new_state.try_into();
-                        match state {
-                            Ok(state) => {
-                                game.set_cell((pos.row() as usize, pos.col() as usize), user_id);
-                                game.set_state(state);
-                                cell_updated.send(CellUpdated::new(*pos, user_id));
-                                state_updated.send(StateUpdated(state));
-                                play_sound(&mut commands, &asset_server, TURN_SOUND_PATH);
-                            }
-                            Err(err) => {
-                                println!("failed to convert game state: {}", err);
-                            }
-                        }
-                    } else {
-                        println!("MakeTurn returned empty response");
-                        play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
-                    }
-                }
-                Err(err) => {
-                    println!("MakeTurn request failed: {}", err);
-                    play_sound(&mut commands, &asset_server, ERROR_SOUND_PATH);
-                }
-            }
-        }
-    }
-}
-
-fn handle_get_game_task(
-    mut get_game: Query<(Entity, &mut CallGetGame)>,
-    mut commands: Commands,
-    mut game: Option<ResMut<CurrentGame>>,
-    mut cell_updated: EventWriter<CellUpdated>,
-    mut state_updated: EventWriter<StateUpdated>,
-) {
-    for (entity, mut task) in get_game.iter_mut() {
-        if let Some(res) = block_on(future::poll_once(&mut task.0)) {
-            commands.entity(entity).remove::<CallGetGame>();
-            let game = match &mut game {
-                Some(game) => game,
-                None => {
-                    println!("no current game, dropping GetGame response");
-                    continue;
-                }
-            };
-            match res {
-                Ok(response) => {
-                    if let Some(info) = response.into_inner().game_info {
-                        if info.game_id != game.id() {
-                            println!("received game info for other game, dropping");
-                            continue;
-                        }
-                        let full_info = match FullGameInfo::try_from(info) {
-                            Ok(info) => info,
-                            Err(err) => {
-                                println!("failed to decode full game info: {}", err);
-                                continue;
-                            }
-                        };
-                        for (i, row) in full_info.board.iter().enumerate() {
-                            for (j, cell) in row.iter().enumerate() {
-                                if let BoardCell(Some(player)) = cell {
-                                    let local_cell = game.board()[i][j];
-                                    if local_cell.is_none() {
-                                        game.set_cell((i, j), *player);
-                                        let pos = Position::new(i as u32, j as u32);
-                                        cell_updated.send(CellUpdated::new(pos, *player));
-                                    }
-                                }
-                            }
-                        }
-                        if full_info.info.state != game.state() {
-                            game.set_state(full_info.info.state);
-                            state_updated.send(StateUpdated(full_info.info.state));
-                        }
-                    } else {
-                        println!("GetGame returned empty response");
-                    }
-                }
-                Err(err) => {
-                    println!("GetGame request failed: {}", err);
-                }
-            }
         }
     }
 }
@@ -869,21 +702,6 @@ fn handle_cell_updated(
             pos: event.pos(),
             image: img.clone(),
         });
-    }
-}
-
-fn handle_state_updated(
-    mut state_updated: EventReader<StateUpdated>,
-    mut game_over: EventWriter<GameOver>,
-) {
-    if let Some(event) = state_updated.read().last() {
-        println!("received StateUpdated: {:?}", event);
-        match event.0 {
-            GameState::Turn(_) => {}
-            GameState::Finished(f) => {
-                game_over.send(GameOver(f));
-            }
-        };
     }
 }
